chore: validate deps
