+++ crates/cli/src/helper/lock.rs
@@ -1,12 +1,15 @@
-use anyhow::{anyhow, Result};
+use anyhow::{anyhow, Result, Context};
 use serde::Deserialize;
 use serde_json::{json, Value};
 use std::path::PathBuf;
 use std::{collections::HashMap, fs};
 
-use crate::util::logger::log_verbose;
+use crate::util::config::get_legacy_peer_deps;
+use crate::util::logger::{log_verbose, log_warning};
+use crate::util::node::Overrides;
 use crate::util::registry::resolve;
 use crate::util::save_type::{PackageAction, SaveType};
+use crate::util::semver;
 use crate::util::{cache::parse_pattern, cloner::clone, downloader::download};
 use crate::{cmd::deps::build_deps, util::logger::log_info};
 
@@ -64,6 +67,11 @@ pub async fn ensure_package_lock() -> Result<()> {
     } else {
         // load package-lock.json directly if exists
         log_info("Loading package-lock.json from current project for dependency download");
+        // Validate dependencies to ensure package-lock.json is in sync with package.json
+        if let Err(_e) = validate_deps().await {
+            log_info("package-lock.json is out of sync with package.json, rebuilding dependencies");
+            build_deps().await?;
+        }
         Ok(())
     }
 }
@@ -284,5 +292,164 @@ mod tests {
             super::path_to_pkg_name("/root/node_modules/@a/b/node_modules/b/c/d"),
             None
         );
+    }
+}
+
+pub async fn validate_deps() -> Result<()> {
+    let path = PathBuf::from(".");
+    let lock_path = path.join("package-lock.json");
+    let pkg_path = path.join("package.json");
+
+    // Read package.json for overrides
+    let pkg_content = fs::read_to_string(pkg_path).context("Failed to read package.json")?;
+    let pkg_file: serde_json::Value = serde_json::from_str(&pkg_content)
+        .map_err(|e| anyhow::anyhow!("Failed to parse package.json: {}", e))?;
+
+    // Initialize overrides
+    let overrides = Overrides::new(pkg_file.clone()).parse(pkg_file);
+
+    let lock_content = fs::read_to_string(lock_path).context("Failed to read package-lock.json")?;
+    let lock_file: serde_json::Value = serde_json::from_str(&lock_content)
+        .map_err(|e| anyhow::anyhow!("Failed to parse package-lock.json: {}", e))?;
+
+    if let Some(packages) = lock_file.get("packages").and_then(|p| p.as_object()) {
+        for (pkg_path, pkg_info) in packages {
+            for (dep_field, is_optional) in get_dep_types() {
+                if let Some(dependencies) = pkg_info.get(dep_field).and_then(|d| d.as_object()) {
+                    for (dep_name, req_version) in dependencies {
+                        let req_version_str = req_version.as_str().unwrap_or_default();
+
+                        // Collect parent chain information
+                        let mut parent_chain = Vec::new();
+                        let mut current_path = if pkg_path.is_empty() {
+                            String::new()
+                        } else {
+                            pkg_path.to_string()
+                        };
+
+                        while !current_path.is_empty() {
+                            if let Some(pkg_info) = packages.get(&current_path) {
+                                if let Some(name) = pkg_info.get("name").and_then(|n| n.as_str()) {
+                                    if let Some(version) =
+                                        pkg_info.get("version").and_then(|v| v.as_str())
+                                    {
+                                        parent_chain.push((name.to_string(), version.to_string()));
+                                    }
+                                }
+                            }
+
+                            if let Some(last_modules) = current_path.rfind("/node_modules/") {
+                                current_path = current_path[..last_modules].to_string();
+                            } else {
+                                current_path = String::new();
+                            }
+                        }
+
+                        // Check if there's an override rule for this dependency
+                        let effective_req_version = if let Some(overrides) = &overrides {
+                            let mut effective_version = req_version_str.to_string();
+                            for rule in &overrides.rules {
+                                if overrides
+                                    .matches_rule(rule, dep_name, req_version_str, &parent_chain)
+                                    .await
+                                {
+                                    effective_version = rule.target_spec.clone();
+                                    break;
+                                }
+                            }
+                            effective_version
+                        } else {
+                            req_version_str.to_string()
+                        };
+
+                        // find the actual version of the dependency
+                        let mut current_path = if pkg_path.is_empty() {
+                            String::new()
+                        } else {
+                            pkg_path.to_string()
+                        };
+                        let mut dep_info = None;
+
+                        // until root or found
+                        loop {
+                            let search_path = if current_path.is_empty() {
+                                format!("node_modules/{}", dep_name)
+                            } else {
+                                format!("{}/node_modules/{}", current_path, dep_name)
+                            };
+
+                            if let Some(info) = packages.get(&search_path) {
+                                dep_info = Some(info);
+                                current_path = search_path;
+                                break;
+                            }
+
+                            // find in root path
+                            if current_path.is_empty() {
+                                break;
+                            }
+
+                            // find in parent path
+                            if let Some(last_modules) = current_path.rfind("/node_modules/") {
+                                current_path = current_path[..last_modules].to_string();
+                            } else {
+                                current_path = String::new();
+                            }
+                        }
+
+                        // optional dependency not found is allowed
+                        if let Some(dep_info) = dep_info {
+                            if let Some(actual_version) =
+                                dep_info.get("version").and_then(|v| v.as_str())
+                            {
+                                println!("effective_req_version: {} actual_version: {}", effective_req_version, actual_version);
+                                if !semver::matches(&effective_req_version, actual_version) {
+                                    log_warning(&format!(
+                                        "Package {} {} dependency {} (required version: {}, effective version: {}) does not match actual version {}@{}",
+                                        pkg_path, dep_field, dep_name, req_version_str, effective_req_version, current_path, actual_version
+                                    ));
+                                    anyhow::bail!("Package {} {} dependency {} (required version: {}, effective version: {}) does not match actual version {}@{}",
+                                        pkg_path, dep_field, dep_name, req_version_str, effective_req_version, current_path, actual_version
+                                    );
+                                }
+                            }
+                        } else if !is_optional {
+                            log_warning(&format!(
+                                "pkg_path {} dep_field {} dep_name {} not found",
+                                pkg_path, dep_field, dep_name
+                            ));
+                            anyhow::bail!(
+                                "Package {} {} dependency {} not found",
+                                pkg_path,
+                                dep_field,
+                                dep_name
+                            );
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    Ok(())
+}
+
+
+fn get_dep_types() -> Vec<(&'static str, bool)> {
+    let legacy_peer_deps = get_legacy_peer_deps();
+
+    if legacy_peer_deps {
+        vec![
+            ("dependencies", false),
+            ("optionalDependencies", true),
+            ("devDependencies", false),
+        ]
+    } else {
+        vec![
+            ("dependencies", false),
+            ("peerDependencies", false),
+            ("optionalDependencies", true),
+            ("devDependencies", false),
+        ]
     }
 }
